Corrected other bug in the ISA of the LEON and ARM processors

+++ processors/LEON3/LEON3Isa.py
@@ -1401,12 +1401,12 @@ opCodeWb = cxx_writer.writer_code.Code("""
 rd = result;
 """)
 opCodeExecS = cxx_writer.writer_code.Code("""
-long long resultTemp = (long long)((int)rs1_op)*((int)rs2_op);
+long long resultTemp = (long long)(((long long)((int)rs1_op))*((long long)((int)rs2_op)));
 Y = ((unsigned long long)resultTemp) >> 32;
 result = resultTemp & 0x00000000FFFFFFFF;
 """)
 opCodeExecU = cxx_writer.writer_code.Code("""
-unsigned long long resultTemp = (unsigned long long)((unsigned int)rs1_op)*((unsigned int)rs2_op);
+unsigned long long resultTemp = (unsigned long long)(((unsigned long long)((unsigned int)rs1_op))*((unsigned long long)((unsigned int)rs2_op)));
 Y = resultTemp >> 32;
 result = resultTemp & 0x00000000FFFFFFFF;
 """)
